Revert "UI fixes"
This reverts commit 515a1d7d37d33bde4626eade6621cd723a300b46.

+++ lib/modules/input/components/ai_overlay.dart
@@ -107,7 +107,7 @@ class AICheckerOverlay extends StatelessWidget {
                         RadialBarSeries<AiDataModel?, String>(
                           dataSource: [data],
                           xValueMapper: (data, _) => "AI Generated",
-                          yValueMapper: (data, _) => data?.fakePercentage as double,
+                          yValueMapper: (data, _) => data?.fakePercentage,
                           maximumValue: 100,
                           trackColor: APColor.light[100]!,
                           cornerStyle: CornerStyle.bothCurve,
